Fixed checking for DeleteCommentPermission

+++ canopeum_backend/canopeum_backend/views.py
@@ -348,6 +348,7 @@ class CommentDetailAPIView(APIView):
         except Comment.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
+        self.check_object_permissions(request, comment)
         comment.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
